templates and url mapping
the account app template have been finished. they follow similar patternes only requiring the change of the header name and the value of the input button.

the account/urls.py done, we used the views that auth has on board for login and logout. and we set the url patterns for the imported onbeard view as a view.. set up the url pattern for the signup page as well.

in the project app url, we inputte the include that tranfers the link to the app.

started off with the base.html found in the project templates folder. and creted the initial linked we wnat to further check.

no testing or debugging done, after i need to further establish the connection.