setting up the Group views and templates
within this push

the views for the pages in the  groups webapp is created with class based views.  These are: createGroup(to create a group), SingleGroup(list of post within a group) and ListGroup(List of all created Views).

Group templates.

1. The group_base.html, i had it extend from the base.html found in the project folder, then further inputted a conted block. however, this content block is much different - this blovk content cotains different keywords that allows to distribute unique code distinct to the webpage, and will be refernced to determine where the certain code should be. this will be displayed further in the all templates within groups.

2. Group_detail Template, this templates displays the use of the different content block being used. within the first content block a title injection and injection of the total members within a group. A check was also inputted in the first content block to give the user the option join or leave.

The the next content block - group_conent.  simply display all the post within the current group. if there isnt a post within the group then it will send an a message that there is no current post with in it. another unique thing added here is the include template tag. there is another  code injection allows the concept of DRY coding