Group_list & Group_form done - not debugeed reupload with description of what's been created
Group_list template,

with in this template the content block division were further displayed.  within the first devision called pregroup. I inputted a check if the user was authenticated. if they were it will give them welcoming greet. with that a linked to their user account page with the list of their created post. lastly display their username next to the greeting message.  if they were not logged in, a simple message welcoming them the groups page.

in the same content block, we did another check if authenticated, so that the user can have an option to create a new group.

The main body content block called - group content. As this is the group_list, we list all the group object and display them individually by name  and their corresponding description. furthermore, a templatetag method was used to count the num of members and post within each forum. this infromation is displayed with each group object being listed.

Group_forum,
this was a more simple task. as it was a form template that derives from the create view, we used bootstrap3 to inject the created form. including the csrf_token. lastly within this page after posting something valid, it will send to the create page.