urls.pu(1/2) done and register the admin
connect the views created groups app to the group app urls. only set up two views connection to the url.

register the group model to the admin.py file , something new  - used a tubular class that  allows the dev to utilies the django admin interface website with the ability to edit models same page of the parent model