Post template
within this push as the as the views and urls has been created. it now time to finish off the majority, if not all the templates that corresonds to the view and urls in the post app.

1. post_base.html
very similar to the group base.html, this will be what all the templates in the post app will inherit. we inputted 3 content block to better structure the layout of the individual page, granted not all the content block will be used in the indivual template but will serve as a greate tool to use.

2.  post_form.html
this is the html template linked to creating a new post.  Very similar to the group_form.html. we have form and input html tag and within these tag we use template tagging to inject the form using bootstrap3 as well as having csrf token to enusre secureity

3. post_list.html.
within this template there is two sections.

The First being the pre_post_content. Within this division contain checks if the user is authenicated and if so then show get all the users groups they joined and display all the groups with a slug format. and each of the listed group will have a link transferring the user to the correspond group. After, if they are't logged in we also display the other groups.

The Second section post_content,
this is more simple, it doesnt require a authenication check but instead looping through all the post and using a special django template tag, we are able to inject a html code file  call '_post.html' to ensure DRY concept.

4. _post.html,
this was called when using the 'include' template tag. This is going to display all the current post message then have the authots username underneath it the message. also have the username link to another page that, list out the users post.  lastly check if the user is authenticated and the current session user is the same post.user then give the option to delete the post.

This is what is being injected when using the call include _post.html

 5. post_confirm_delete.html,
This is a simpleir task as the _post.html has been discussed. this template will initially show the current _post.html object and give the option to confirm the deletion or cancel which will send the user back to the original post.

6. post_detail.html
an even more simple task, as the majority of the script was done in _post.html which is basically a post detail. all this template contains is a include template tag that injects _post.html

7. user_post_list.html
within this template it is split into 2 divisions however it is very simple
the first division called prepost displays the current users username and the second division list all the user post using the _post.html injection call.

IMPORTANT - this will contain bugs!! next session will deal with setting up the setting to start debuging.